+++ examples-m/main.js
@@ -7,8 +7,27 @@ import './assets/style/base.scss'
 import WhiteSpace from './components/whiteSpace.vue'
 import Space from './components/space.vue'
 import '@vant/doc/helper/touch-simulator'
+import VueI18n from 'vue-i18n'
+import zhCN from '@/locale/lang/zh-CN'
+import enUS from '@/locale/lang/en-US'
 
-Vue.use(JMDesign)
+Vue.use(VueI18n)
+const i18n = new VueI18n({
+  locale: 'zh_CN',
+  messages: {
+    'zh_CN': {
+      hello: '你好',
+      ...zhCN
+    },
+    'en_US': {
+      hello: 'Hello',
+      ...enUS
+    }
+  }
+})
+Vue.use(JMDesign, {
+  i18n: (key, value) => i18n.t(key, value)
+})
 Vue.component(WhiteSpace.name, WhiteSpace)
 Vue.component(Space.name, Space)
 Vue.config.productionTip = false
@@ -17,6 +36,7 @@ Vue.config.productionTip = false
 new Vue({
   el: '#app',
   router,
+  i18n,
   template: '<App/>',
   components: { App }
 })