+++ src/mixins/popup/index.js
@@ -1,7 +1,26 @@
 import { context } from './context'
 import { openModal, closeModal } from './modal'
+import TouchMixin from '../touch'
+import { getScrollEventTarget } from '../../utils/scroll'
+import { on, off } from '../../utils/event'
+
+export function stopPropagation (event) {
+  event.stopPropagation()
+}
+
+export function preventDefault (event, isStopPropagation) {
+  /* istanbul ignore else */
+  if (typeof event.cancelable !== 'boolean' || event.cancelable) {
+    event.preventDefault()
+  }
+
+  if (isStopPropagation) {
+    stopPropagation(event)
+  }
+}
 
 export default {
+  mixins: [TouchMixin],
   data () {
     return {
       inited: this.value
@@ -16,6 +35,11 @@ export default {
     closeOnClickModal: {
       type: Boolean,
       default: true
+    },
+    zIndex: Number,
+    duration: {
+      type: Number,
+      default: 300
     }
   },
   watch: {
@@ -28,11 +52,19 @@ export default {
   },
   methods: {
     open () {
-      if (this.$isServer) return
+      if (this.$isServer || this.opened) return
 
+      if (this.zIndex !== undefined) {
+        context.zIndex = this.zIndex
+      }
+
+      this.opened = true
       this.renderModal()
 
       if (this.lockScroll) {
+        on(document, 'touchstart', this.touchStart)
+        on(document, 'touchmove', this.onTouchMove)
+
         if (!context.lockCount) {
           document.body.classList.add('jm-overflow-hidden')
         }
@@ -40,7 +72,11 @@ export default {
       }
     },
     close () {
+      if (!this.opened) return
+
       if (this.lockScroll) {
+        off(document, 'touchstart', this.touchStart)
+        off(document, 'touchmove', this.onTouchMove)
         context.lockCount--
 
         if (!context.lockCount) {
@@ -48,6 +84,7 @@ export default {
         }
       }
 
+      this.opened = false
       closeModal(this)
       this.$emit('input', false)
     },
@@ -58,6 +95,25 @@ export default {
         zIndex: context.zIndex++,
         closeOnClickModal: this.closeOnClickModal
       })
+    },
+    onTouchMove (event) {
+      this.touchMove(event)
+      const direction = this.deltaY > 0 ? '10' : '01'
+      const el = getScrollEventTarget(event.target, this.$el)
+      const { scrollHeight, offsetHeight, scrollTop } = el
+      let status = '11'
+
+      if (scrollTop === 0) {
+        status = offsetHeight >= scrollHeight ? '00' : '01'
+      } else if (scrollTop + offsetHeight >= scrollHeight) {
+        status = '10'
+      }
+
+      if (status !== '11' &&
+        this.direction === 'vertical' &&
+        !(parseInt(status, 2) & parseInt(direction, 2))) {
+        preventDefault(event)
+      }
     }
   },
   mounted () {